requirements.txt
playwright==1.48.0
playwright-stealth==1.0.6
fastapi==0.115.0
uvicorn[standard]==0.32.0
typer==0.15.1
pydantic==2.9.2
pydantic-settings==2.6.0
sqlalchemy==2.0.36
aiosqlite==0.20.0
python-multipart==0.0.12
python-jose[cryptography]==3.3.0
PyJWT==2.9.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
httpx==0.27.2
slowapi==0.1.9
websocket-client==1.8.0
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-cov==6.0.0
black==24.8.0
isort==5.13.2
flake8==7.1.1
mypy==1.13.0
psutil==6.1.0
dearpygui==1.14.1
